Number of Islands
The numIslands method takes a 2D array of characters grid as input and returns the count of islands present in the grid.
Inside the numIslands method:
A boolean array visited is initialized to mark visited cells. Initially, all cells are unvisited.
Two nested loops iterate through all cells of the given matrix (grid).
If a cell with value '1' is encountered and it has not been visited yet (!visited[i][j]), it indicates the start of a new island. The Depth-First Search (DFS) algorithm is then invoked to explore and mark all cells belonging to this island.
Each time a new island is encountered, the count variable is incremented.
The DFS method is a recursive helper function used to explore an island starting from a given cell (i, j). It traverses all adjacent cells of the current cell (i, j) and marks them as visited if they belong to the same island. This process continues recursively until all cells of the island are visited.
Finally, the count variable, which represents the total number of islands in the grid, is returned as the result.